Fix tests and rename namespace
