+++ library/Vanilla/EmbeddedContent/Factories/DiscussionEmbedFactory.php
@@ -5,12 +5,13 @@
  * @license GPL-2.0-only
  */
 
-namespace Vanilla\EmbeddedContent\Embeds;
+namespace Vanilla\EmbeddedContent\Factories;
 
 use Garden\Web\Exception\NotFoundException;
 use Garden\Web\RequestInterface;
 use Vanilla\EmbeddedContent\AbstractEmbed;
 use Vanilla\EmbeddedContent\AbstractEmbedFactory;
+use Vanilla\EmbeddedContent\Embeds\QuoteEmbed;
 use Vanilla\Web\Asset\SiteAsset;
 
 /**
