+++ library/Vanilla/EmbeddedContent/Factories/ScrapeEmbedFactory.php
@@ -4,11 +4,12 @@
  * @license GPL-2.0-only
  */
 
-namespace Vanilla\EmbeddedContent\Embeds;
+namespace Vanilla\EmbeddedContent\Factories;
 
 use Garden\Http\HttpClient;
 use Vanilla\EmbeddedContent\AbstractEmbed;
-use Vanilla\EmbeddedContent\AbstractEmbedFactory;
+use Vanilla\EmbeddedContent\Embeds\ImageEmbed;
+use Vanilla\EmbeddedContent\Embeds\LinkEmbed;
 use Vanilla\EmbeddedContent\FallbackEmbedFactory;
 use Vanilla\PageScraper;
