+++ tests/Library/Vanilla/EmbeddedContent/Embeds/QuoteEmbedTest.php
@@ -10,12 +10,24 @@
 namespace VanillaTests\Library\EmbeddedContent\Embeds;
 
 use Vanilla\EmbeddedContent\Embeds\QuoteEmbed;
+use Vanilla\Formatting\Formats\RichFormat;
+use Vanilla\Formatting\FormatService;
 use VanillaTests\ContainerTestCase;
 
 /**
  * Test for the individual linkembed.
  */
 class QuoteEmbedTest extends ContainerTestCase {
+
+    /**
+     * Setup.
+     */
+    public function setUp() {
+        parent::setUp();
+        $this->container->rule(FormatService::class)
+            ->addCall('registerFormat', [RichFormat::FORMAT_KEY, RichFormat::class]);
+    }
+
     /**
      * Ensure we can create discussion embed from the old data format that might still
      * live in the DB.
@@ -33,37 +45,7 @@ class QuoteEmbedTest extends ContainerTestCase {
     "attributes": {
         "discussionID": 8,
         "name": "test file upload",
-        "bodyRaw": [
-            {
-                "insert": {
-                    "embed-external": {
-                        "data": {
-                            "url": "https://dev.vanilla.localhost/uploads/150/LKE0S2FWLFUP.zip",
-                            "type": "file",
-                            "attributes": {
-                                "mediaID": 62,
-                                "name": "___img_onload_prompt(1)_ (2).zip",
-                                "path": "150/LKE0S2FWLFUP.zip",
-                                "type": "application/zip",
-                                "size": 41,
-                                "active": 1,
-                                "insertUserID": 4,
-                                "dateInserted": "2019-06-14 14:09:38",
-                                "foreignID": 4,
-                                "foreignTable": "embed",
-                                "imageWidth": null,
-                                "imageHeight": null,
-                                "thumbWidth": null,
-                                "thumbHeight": null,
-                                "thumbPath": null,
-                                "foreignType": "embed",
-                                "url": "https://dev.vanilla.localhost/uploads/150/LKE0S2FWLFUP.zip"
-                            }
-                        }
-                    }
-                }
-            }
-        ],
+        "bodyRaw": [{ "insert": "Testtes test\\n" }],
         "dateInserted": "2019-06-14T14:09:45+00:00",
         "dateUpdated": null,
         "insertUser": {
