+++ tests/SharedBootstrapTestCase.php
@@ -18,9 +18,8 @@ use Garden\Container\Container;
  */
 class SharedBootstrapTestCase extends TestCase {
 
-    use BootstrapTrait {
-        setupBeforeClass as private bootstrapBeforeClass;
-    }
+    use BootstrapTrait;
+
     /**
      * Bootstrap the first test cases and reuse the same container/bootstrap for subsequent test cases.
      */
@@ -34,7 +33,7 @@ class SharedBootstrapTestCase extends TestCase {
             if (!self::containerIsNull($currentContainer)) {
                 $containerCorruption = true;
             } else {
-                self::bootstrapBeforeClass();
+                BootstrapTrait::setUpBeforeClass();
             }
         } else {
             if (!self::containerIsNull($currentContainer) && $currentContainer !== self::$container) {
@@ -65,5 +64,12 @@ class SharedBootstrapTestCase extends TestCase {
      */
     protected static function getBootstrapFolderName() {
         return 'sharedbootstrap';
+    }
+
+    /**
+     * Cleanup the container after testing is done.
+     */
+    public static function tearDownAfterClass() {
+        Bootstrap::cleanUpGlobals();
     }
 }
